Rename and add performance optimisations
