+++ merge_utils/__init__.py
@@ -5,6 +5,6 @@ This package provides tools for partition-aware merge operations
 with configurable update behaviors, similar to Delta Lake MERGE operations.
 """
 
-from utils.merge_generator import get_merged_df
+from merge_utils.merge_generator import get_merged_df
 
 __all__ = ['get_merged_df']
